# Produce correct answers for Group BY NULL (Option 1)