@alamb 2021-07-29
Clippy error is unrelated -- see fix in https://github.com/apache/arrow-datafusion/pull/794