@jhorstmann 2021-07-30
Looks good. I was trying to come up with an example where two distinct keys would end up as the same encoding but could not find any because any variable length types include the length prefix. :rocket: