@alamb 2021-07-30
Thanks for the reviews @Dandandan  and @jhorstmann ! I plan to wait another day or two to see if anyone else has feedback on this approach, but what I am thinking of doing is merging this PR (after addressing comments) so at least we get correct answers and then work on the more sophisticated implementation in parallel. 

> Looks good. I was trying to come up with an example where two distinct keys would end up as the same encoding but could not find any because any variable length types include the length prefix. 🚀

Thanks for double checking -- this also worried me a lot so I am glad to hear someone else did a double check too. 

I convinced myself that since each key had entries for the same columns in the same order, there was no way to concoct the same bytes with different column values