@alamb 2021-07-30
Rebased given that https://github.com/apache/arrow-datafusion/pull/786 is merged, so that this PR just shows the delta now