@alamb 2021-08-02
https://github.com/apache/arrow-datafusion/pull/808 contains the PR that should give us back any performance we lost in this one